Add undoIgnore to selection restoration editEntityRecord calls
Selection restoration on entity navigation should not create undo
levels, as it's an internal mechanism, not a user edit. Also moves
selection restoration logic into useResolveEditedEntity for cleaner
separation of concerns.
